# backend/Fms/openapi/index.yaml
openapi: 3.0.1
info:
  title: FMS
  version: v1
servers:
  - url: http://localhost:3333
paths:
...
  '/api/workspaces/{workspaceId}/transactions':
    put:
      tags:
        - Transaction
      summary: Create new transaction
      operationId: createTransaction
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpsertRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/TransactionUpsertRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/TransactionUpsertRequest'
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: integer
                format: int32
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    get:
      tags:
        - Transaction
      summary: Get list of workspace transactions
      operationId: listTransactions
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: CategoryId
          in: query
          schema:
            type: integer
            format: int32
            example: 35
          example: 35
        - name: StartDate
          in: query
          schema:
            type: string
            format: date-time
            example: '2024-06-07T07:00:00.000Z'
          example: '2024-06-07T07:00:00.000Z'
        - name: EndDate
          in: query
          schema:
            type: string
            format: date-time
            example: 
          example: 'null'
        - name: UserId
          in: query
          schema:
            type: integer
            format: int32
            example: 
          example: 
        - name: MinAmount
          in: query
          schema:
            type: integer
            format: int32
            example: 300
          example: 300
        - name: MaxAmount
          in: query
          schema:
            type: integer
            format: int32
            example: 
          example: 
        - name: IncludeHistory
          in: query
          schema:
            type: boolean
            example: 
          example: 
        - name: CategoryKind
          in: query
          schema:
            $ref: '#/components/schemas/TransactionCategoryKind'
          example: 'null'
        - name: SortField
          in: query
          schema:
            $ref: '#/components/schemas/TransactionSortField'
        - name: SortDirection
          in: query
          schema:
            $ref: '#/components/schemas/SortDirection'
        - name: Offset
          in: query
          schema:
            type: integer
            format: int32
        - name: Limit
          in: query
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
  '/api/transactions/{id}':
    get:
      tags:
        - Transaction
      summary: Get transaction
      operationId: getTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    post:
      tags:
        - Transaction
      summary: Update transaction
      operationId: updateTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpsertRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/TransactionUpsertRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/TransactionUpsertRequest'
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    delete:
      tags:
        - Transaction
      summary: Delete transaction
      operationId: deleteTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
  '/api/workspaces/{workspaceId}/transactions/group-by-category':
    get:
      tags:
        - Transaction
      summary: Get list of workspace transactions grouped by category
      operationId: listTransactionsGroupByCategory
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: CategoryId
          in: query
          schema:
            type: integer
            format: int32
            example: 35
          example: 35
        - name: StartDate
          in: query
          schema:
            type: string
            format: date-time
            example: '2024-06-07T07:00:00.000Z'
          example: '2024-06-07T07:00:00.000Z'
        - name: EndDate
          in: query
          schema:
            type: string
            format: date-time
            example: 
          example: 'null'
        - name: UserId
          in: query
          schema:
            type: integer
            format: int32
            example: 
          example: 
        - name: MinAmount
          in: query
          schema:
            type: integer
            format: int32
            example: 300
          example: 300
        - name: MaxAmount
          in: query
          schema:
            type: integer
            format: int32
            example: 
          example: 
        - name: IncludeHistory
          in: query
          schema:
            type: boolean
            example: 
          example: 
        - name: CategoryKind
          in: query
          schema:
            $ref: '#/components/schemas/TransactionCategoryKind'
          example: 'null'
        - name: SortField
          in: query
          schema:
            $ref: '#/components/schemas/TransactionSortField'
        - name: SortDirection
          in: query
          schema:
            $ref: '#/components/schemas/SortDirection'
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionGroupedByCategoryListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
  '/api/workspaces/{workspaceId}/transactions/group-by-user':
    get:
      tags:
        - Transaction
      summary: Get list of workspace transactions grouped by user
      operationId: listTransactionsGroupByUser
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: CategoryId
          in: query
          schema:
            type: integer
            format: int32
            example: 35
          example: 35
        - name: StartDate
          in: query
          schema:
            type: string
            format: date-time
            example: '2024-06-07T07:00:00.000Z'
          example: '2024-06-07T07:00:00.000Z'
        - name: EndDate
          in: query
          schema:
            type: string
            format: date-time
            example: 
          example: 'null'
        - name: UserId
          in: query
          schema:
            type: integer
            format: int32
            example: 
          example: 
        - name: MinAmount
          in: query
          schema:
            type: integer
            format: int32
            example: 300
          example: 300
        - name: MaxAmount
          in: query
          schema:
            type: integer
            format: int32
            example: 
          example: 
        - name: IncludeHistory
          in: query
          schema:
            type: boolean
            example: 
          example: 
        - name: CategoryKind
          in: query
          schema:
            $ref: '#/components/schemas/TransactionCategoryKind'
          example: 'null'
        - name: SortField
          in: query
          schema:
            $ref: '#/components/schemas/TransactionSortField'
        - name: SortDirection
          in: query
          schema:
            $ref: '#/components/schemas/SortDirection'
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionGroupedByUserListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
...
components:
  schemas:
...
    AccountResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        organization:
          $ref: '#/components/schemas/OrganizationShortResponse'
      additionalProperties: false
...
    OrganizationShortResponse:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 243
        name:
          minLength: 1
          type: string
          example: Cool Organization Inc.
      additionalProperties: false
...
    PublicClientError:
      required:
        - description
      type: object
      properties:
        description:
          minLength: 1
          type: string
          example: Something bad happened
        validationErrors:
          type: array
          items:
            type: string
          nullable: true
          example:
            - Email can't contain less than 3 characters
            - Password can't be empty
      additionalProperties: false
    PublicError:
      required:
        - description
      type: object
      properties:
        description:
          minLength: 1
          type: string
          example: Something bad happened
      additionalProperties: false
    SortDirection:
      enum:
        - ASCENDING
        - DESCENDING
      type: string
...
    TransactionCategoryKind:
      enum:
        - INCOME
        - EXPENSE
        - MIXED
      type: string
...
    TransactionCategoryResponse:
      required:
        - id
        - kind
        - name
        - ownerAccount
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 5359
        name:
          minLength: 1
          type: string
          example: Salary
        ownerAccount:
          $ref: '#/components/schemas/AccountResponse'
        kind:
          $ref: '#/components/schemas/TransactionCategoryKind'
        uiColor:
          type: string
          nullable: true
          example: FF0022
        workspace:
          $ref: '#/components/schemas/WorkspaceShortResponse'
      additionalProperties: false
...
    TransactionGroupedByCategoryListResponse:
      required:
        - items
        - totalAmount
      type: object
      properties:
        totalAmount:
          type: integer
          format: int32
          example: -340
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionGroupedByCategoryResponse'
      additionalProperties: false
    TransactionGroupedByCategoryResponse:
      required:
        - amount
        - category
      type: object
      properties:
        category:
          $ref: '#/components/schemas/TransactionCategoryResponse'
        history:
          type: array
          items:
            $ref: '#/components/schemas/TransactionShortResponse'
          nullable: true
        amount:
          type: integer
          format: int32
          example: 15320
      additionalProperties: false
    TransactionGroupedByUserListResponse:
      required:
        - items
        - totalAmount
      type: object
      properties:
        totalAmount:
          type: integer
          format: int32
          example: 0
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionGroupedByUserResponse'
      additionalProperties: false
    TransactionGroupedByUserResponse:
      required:
        - amount
        - user
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        history:
          type: array
          items:
            $ref: '#/components/schemas/TransactionShortResponse'
          nullable: true
        amount:
          type: integer
          format: int32
          example: -4320
      additionalProperties: false
    TransactionListResponse:
      required:
        - items
        - totalCount
      type: object
      properties:
        totalCount:
          type: integer
          format: int32
          example: 165
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
      additionalProperties: false
    TransactionResponse:
      required:
        - amount
        - category
        - creationTimestamp
        - creationUser
        - id
        - timestamp
        - workspace
      type: object
      properties:
        id:
          type: integer
          format: int32
        workspace:
          $ref: '#/components/schemas/WorkspaceShortResponse'
        category:
          $ref: '#/components/schemas/TransactionCategoryResponse'
        amount:
          type: integer
          format: int32
          example: -530
        description:
          type: string
          nullable: true
          example: Some optional description
        timestamp:
          type: string
          format: date-time
        creationTimestamp:
          type: string
          format: date-time
        creationUser:
          $ref: '#/components/schemas/UserResponse'
        user:
          $ref: '#/components/schemas/UserResponse'
      additionalProperties: false
    TransactionShortResponse:
      required:
        - amount
        - id
        - timestamp
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 2138
        amount:
          type: integer
          format: int32
          example: 120
        timestamp:
          type: string
          format: date-time
      additionalProperties: false
    TransactionSortField:
      enum:
        - TIMESTAMP
        - AMOUNT
      type: string
    TransactionUpsertRequest:
      required:
        - amount
        - categoryId
      type: object
      properties:
        categoryId:
          type: integer
          format: int32
          example: 6402
        amount:
          type: integer
          format: int32
          example: 640
        description:
          maxLength: 2048
          type: string
          nullable: true
          example: Some different optional description
        timestamp:
          type: string
          format: date-time
          nullable: true
          example: 
        userId:
          type: integer
          format: int32
          nullable: true
          example: 
      additionalProperties: false
...
    UserResponse:
      required:
        - firstName
        - id
        - lastName
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 25
        firstName:
          minLength: 1
          type: string
          example: John
        lastName:
          minLength: 1
          type: string
          example: Snow
      additionalProperties: false
...
    WorkspaceKind:
      enum:
        - PRIVATE
        - SHARED
      type: string
...
    WorkspaceShortResponse:
      required:
        - id
        - kind
        - name
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 1402
        name:
          minLength: 1
          type: string
          example: Family Workspace
        kind:
          $ref: '#/components/schemas/WorkspaceKind'
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: "JWT Authorization header using the Bearer scheme.\n\nEnter 'Bearer' [space] and then your token in the text input below.\nExample: 'Bearer 12345abcdef'"
      name: Authorization
      in: header
security:
  - Bearer: [ ]